+++ quax/__init__.py
@@ -1,9 +1,6 @@
 from . import integrals 
 from . import constants
 
-if constants.libint_imported:
-    from . import external_integrals 
-
 from . import methods 
 from . import core
 from . import utils
