--- quax/external_integrals/__init__.py
@@ -1,9 +1,0 @@
-from . import tei
-from . import oei
-from . import libint_interface
-
-from .tei import TEI 
-from .oei import OEI 
-
-from .tmp_potential import tmp_potential
-
