+++ quax/integrals/makefile
@@ -2,22 +2,22 @@
 # Eigen headers, Python headers, Pybind11 headers, Libint API headers libint2.h libint2.hpp, the rest of the Libint2 headers, and the library location of libint2.a,
 CC      := g++
 # Libint prefix location (where /include, /include/libint2, /lib, /share are located) 
-LIBINT_PREFIX := /home/vulcan/ecm23353/Code/bin/libint-2.8.0/PREFIX
+LIBINT_PREFIX := /home/ecm23353/psi_env
 
 I1 := $(LIBINT_PREFIX)/include
 I2 := $(LIBINT_PREFIX)/include/libint2
 L1 := $(LIBINT_PREFIX)/lib
 # Eigen headers location 
-I3 := /home/vulcan/ecm23353/.conda/envs/quax/include/eigen3
+I3 := /home/ecm23353/psi_env/include/eigen3
 # Python headers location 
-I4 := /home/vulcan/ecm23353/.conda/envs/quax/include/python3.10
+I4 := /home/ecm23353/psi_env/include/python3.10
 # Pybind11 headers location 
-I5 := /home/vulcan/ecm23353/.conda/envs/quax/lib/python3.10/site-packages/pybind11/include
+I5 := /home/ecm23353/psi_env/lib/python3.10/site-packages/pybind11/include
 # HDF5 headers, static and shared libraries 
-I6 := /home/vulcan/ecm23353/.conda/envs/quax/include
-L2 := /home/vulcan/ecm23353/.conda/envs/quax/lib
+I6 := /home/ecm23353/psi_env/include
+L2 := /home/ecm23353/psi_env/lib
 # Edit path in quotes to be same location as L2 definition above
-RPATH := -Wl,-rpath,"/home/vulcan/ecm23353/.conda/envs/quax/lib"
+RPATH := -Wl,-rpath,"/home/ecm23353/psi_env/lib"
 
 # This 'TARGETS' suffix should be set to whatever is returned by the command `python3-config --extension-suffix` entered on command line.
 # and it should match the same python version referenced in the above include path for I4 := (3.7 in this case)
@@ -33,7 +33,7 @@ clean:
 	rm -f $(OBJ)
 
 $(OBJ): %.o : %.cc $(DEPS)
-	$(CC) -c $< -o $@ -O3 -fPIC -std=c++11 -I $(I1) -I $(I2) -I $(I3) -I $(I4) -I $(I5) -I $(I6) -lint2 -L $(L1) -lhdf5 -lhdf5_cpp -L $(L2) $(RPATH)
+	$(CC) -c $< -o $@ -O3 -fPIC -I $(I1) -I $(I2) -I $(I3) -I $(I4) -I $(I5) -I $(I6) -lint2 -L $(L1) -lhdf5 -lhdf5_cpp -L $(L2) $(RPATH)
 $(TARGETS): $(OBJ)
-	$(CC) $^ -o $@ -O3 -fPIC -shared -std=c++11 -I $(I1) -I $(I2) -I $(I3) -I $(I4) -I $(I5) -I $(I6) -lint2 -L $(L1) -lhdf5 -lhdf5_cpp -L $(L2) $(RPATH)
+	$(CC) $^ -o $@ -O3 -fPIC -shared -I $(I1) -I $(I2) -I $(I3) -I $(I4) -I $(I5) -I $(I6) -lint2 -L $(L1) -lhdf5 -lhdf5_cpp -L $(L2) $(RPATH)
