+++ quax/integrals/oei.py
@@ -1,225 +1,249 @@
-import jax 
-from jax.config import config; config.update("jax_enable_x64", True)
+import jax
 import jax.numpy as jnp
 import numpy as np
-from jax.lax import fori_loop, while_loop
-from functools import partial
+import h5py
+import os
+import psi4
+from . import libint_interface
+from ..utils import get_deriv_vec_idx, how_many_derivs
 
-from .integrals_utils import gaussian_product, boys, binomial_prefactor, factorials, double_factorials, neg_one_pow, cartesian_product, am_leading_indices, angular_momentum_combinations
-from .basis_utils import flatten_basis_data, get_nbf
+jax.config.update("jax_enable_x64", True)
 
-# Useful resources: Fundamentals of Molecular Integrals Evaluation, Fermann, Valeev https://arxiv.org/abs/2007.12057
-#                   Gaussian-Expansion methods of molecular integrals Taketa, Huzinaga, O-ohata 
+class OEI(object):
 
-def overlap(la,ma,na,lb,mb,nb,aa,bb,PA_pow,PB_pow,prefactor):
-    """
-    Computes a single overlap integral. Taketa, Huzinaga, Oohata 2.12
-    P = gaussian product of aa,A; bb,B
-    PA_pow, PB_pow
-        All powers of Pi-Ai or Pi-Bi packed into an array
-        [[(Px-Ax)^0, (Px-Ax)^1, ... (Px-Ax)^max_am]
-         [(Py-Ay)^0, (Py-Ay)^1, ... (Py-Ay)^max_am]
-         [(Pz-Az)^0, (Pz-Az)^1, ... (Pz-Az)^max_am]]
-    prefactor = jnp.exp(-aa * bb * jnp.dot(A-B,A-B) / gamma)
-    """
-    gamma = aa + bb
-    prefactor *= (jnp.pi / gamma)**1.5
+    def __init__(self, basis_name, xyz_path, max_deriv_order, mode):
+        with open(xyz_path, 'r') as f:
+            tmp = f.read()
+        molecule = psi4.core.Molecule.from_string(tmp, 'xyz+')
+        basis_set = psi4.core.BasisSet.build(molecule, 'BASIS', basis_name, puream=0)
+        natoms = molecule.natom()
+        nbf = basis_set.nbf()
 
-    wx = overlap_component(la,lb,PA_pow[0],PB_pow[0],gamma)
-    wy = overlap_component(ma,mb,PA_pow[1],PB_pow[1],gamma)
-    wz = overlap_component(na,nb,PA_pow[2],PB_pow[2],gamma)
-    return prefactor * wx * wy * wz
+        # TODO implement core-algo for OEI's in libint_interface.cc
+        #if mode == 'core' and max_deriv_order > 0:
+            #self.oei_derivatives = {}
 
-def overlap_component(l1,l2,PAx,PBx,gamma):
-    """
-    The 1d overlap integral component. Taketa, Huzinaga, Oohata 2.12
-    """
-    K = 1 + (l1 + l2) // 2  
+        self.mode = mode
+        self.nbf = nbf
 
-    def loop_i(arr):
-       i, total = arr
-       return (i+1, total + binomial_prefactor(2*i,l1,l2,PAx,PBx) * double_factorials[2*i-1] / (2*gamma)**i)
+        # Create new JAX primitives for overlap, kinetic, potential evaluation and their derivatives
+        self.overlap_p = jax.core.Primitive("overlap")
+        self.overlap_deriv_p = jax.core.Primitive("overlap_deriv")
+        self.kinetic_p = jax.core.Primitive("kinetic")
+        self.kinetic_deriv_p = jax.core.Primitive("kinetic_deriv")
+        self.potential_p = jax.core.Primitive("potential")
+        self.potential_deriv_p = jax.core.Primitive("potential_deriv")
 
-    i_accu, total_sum = while_loop(lambda arr: arr[0] < K, loop_i, (0, 0)) # (i, total)
+        # Register primitive evaluation rules
+        self.overlap_p.def_impl(self.overlap_impl)
+        self.overlap_deriv_p.def_impl(self.overlap_deriv_impl)
+        self.kinetic_p.def_impl(self.kinetic_impl)
+        self.kinetic_deriv_p.def_impl(self.kinetic_deriv_impl)
+        self.potential_p.def_impl(self.potential_impl)
+        self.potential_deriv_p.def_impl(self.potential_deriv_impl)
 
-    return total_sum
+        # Register the JVP rules with JAX
+        jax.interpreters.ad.primitive_jvps[self.overlap_p] = self.overlap_jvp
+        jax.interpreters.ad.primitive_jvps[self.overlap_deriv_p] = self.overlap_deriv_jvp
+        jax.interpreters.ad.primitive_jvps[self.kinetic_p] = self.kinetic_jvp
+        jax.interpreters.ad.primitive_jvps[self.kinetic_deriv_p] = self.kinetic_deriv_jvp
+        jax.interpreters.ad.primitive_jvps[self.potential_p] = self.potential_jvp
+        jax.interpreters.ad.primitive_jvps[self.potential_deriv_p] = self.potential_deriv_jvp
 
-def kinetic(la,ma,na,lb,mb,nb,aa,bb,PA_pow,PB_pow,prefactor):
-    """
-    Computes a single kinetic energy integral.
-    """
-    gamma = aa + bb
-    prefactor *= (jnp.pi / gamma)**1.5
-    wx = overlap_component(la,lb,PA_pow[0],PB_pow[0],gamma)
-    wy = overlap_component(ma,mb,PA_pow[1],PB_pow[1],gamma)
-    wz = overlap_component(na,nb,PA_pow[2],PB_pow[2],gamma)
-    wx_plus2 = overlap_component(la,lb+2,PA_pow[0],PB_pow[0],gamma)
-    wy_plus2 = overlap_component(ma,mb+2,PA_pow[1],PB_pow[1],gamma)
-    wz_plus2 = overlap_component(na,nb+2,PA_pow[2],PB_pow[2],gamma)
-    wx_minus2 = overlap_component(la,lb-2,PA_pow[0],PB_pow[0],gamma)
-    wy_minus2 = overlap_component(ma,mb-2,PA_pow[1],PB_pow[1],gamma)
-    wz_minus2 = overlap_component(na,nb-2,PA_pow[2],PB_pow[2],gamma)
+        # Register the batching rules with JAX
+        jax.interpreters.batching.primitive_batchers[self.overlap_deriv_p] = self.overlap_deriv_batch
+        jax.interpreters.batching.primitive_batchers[self.kinetic_deriv_p] = self.kinetic_deriv_batch
+        jax.interpreters.batching.primitive_batchers[self.potential_deriv_p] = self.potential_deriv_batch
 
-    term1 = bb*(2*(lb+mb+nb)+3) * wx * wy * wz 
+    # Create functions to call primitives
+    def overlap(self, geom):
+        return self.overlap_p.bind(geom)
 
-    term2 = -2 * bb**2 * (wx_plus2*wy*wz + wx*wy_plus2*wz + wx*wy*wz_plus2)
+    def overlap_deriv(self, geom, deriv_vec):
+        return self.overlap_deriv_p.bind(geom, deriv_vec)
 
-    term3 = -0.5 * (lb * (lb-1) * wx_minus2 * wy * wz \
-                  + mb * (mb-1) * wx * wy_minus2 * wz \
-                  + nb * (nb-1) * wx * wy * wz_minus2)
-    return prefactor * (term1 + term2 + term3)
+    def kinetic(self, geom):
+        return self.kinetic_p.bind(geom)
 
-def A_array(l1,l2,PA,PB,CP,g,A_vals):
+    def kinetic_deriv(self, geom, deriv_vec):
+        return self.kinetic_deriv_p.bind(geom, deriv_vec)
 
-    def loop_i(arr0):
-       i_0, r_0, u_0, A_0 = arr0
-       Aterm_0 = neg_one_pow[i_0] * binomial_prefactor(i_0,l1,l2,PA,PB) * factorials[i_0]
-       r_0 = i_0 // 2
+    def potential(self, geom):
+        return self.potential_p.bind(geom)
 
-       def loop_r(arr1):
-          i_1, r_1, u_1, Aterm_1, A_1 = arr1
-          u_1 = (i_1 - 2 * r_1) // 2
+    def potential_deriv(self, geom, deriv_vec):
+        return self.potential_deriv_p.bind(geom, deriv_vec)
 
-          def loop_u(arr2):
-             i_2, r_2, u_2, Aterm_2, A_2 = arr2
-             I = i_2 - 2 * r_2 - u_2
-             tmp = I - u_2
-             fact_ratio = 1 / (factorials[r_2] * factorials[u_2] * factorials[tmp])
-             Aterm_2 *= neg_one_pow[u_2]  * CP[tmp] * (0.25 / g)**(r_2+u_2) * fact_ratio
-             A_2 = A_2.at[I].set(Aterm_2)
-             u_2 -= 1
-             return (i_2, r_2, u_2, Aterm_2, A_2)
+    # Create primitive evaluation rules
+    def overlap_impl(self, geom):
+        S = libint_interface.overlap()
+        S = S.reshape(self.nbf,self.nbf)
+        return jnp.asarray(S)
 
-          i_1_, r_1_, u_1_, Aterm_1_, A_1_ = while_loop(lambda arr2: arr2[1] > -1, loop_u, (i_1, r_1, u_1, Aterm_1, A_1))
-          r_1_ -= 1
-          return (i_1_, r_1_, u_1_, Aterm_1_, A_1_)
+    def kinetic_impl(self, geom):
+        T = libint_interface.kinetic()
+        T = T.reshape(self.nbf,self.nbf)
+        return jnp.asarray(T)
 
-       i_0_, r_0_, u_0_, Aterm_0_, A_0_ = while_loop(lambda arr1: arr1[1] > -1, loop_r, (i_0, r_0, u_0, Aterm_0, A_0))
-       i_0_ -= 1
-       return (i_0_, r_0_, u_0_, A_0_)
+    def potential_impl(self, geom):
+        V = libint_interface.potential()
+        V = V.reshape(self.nbf,self.nbf)
+        return jnp.asarray(V)
 
-    i, r, u, A = while_loop(lambda arr0: arr0[0] > -1, loop_i, (l1 + l2, 0, 0, A_vals)) # (i, r, u, A)
+    def overlap_deriv_impl(self, geom, deriv_vec):
+        deriv_vec = np.asarray(deriv_vec, int)
+        deriv_order = np.sum(deriv_vec)
 
-    return A
+        #TODO update once core algo in libint is computed, this just computes one slice at a time
+        if self.mode == 'core':
+            S = libint_interface.overlap_deriv(np.asarray(deriv_vec, int))
+            return jnp.asarray(S).reshape(self.nbf,self.nbf)
+        else:
+            idx = get_deriv_vec_idx(deriv_vec)
+            if os.path.exists("oei_derivs.h5"):
+                file_name = "oei_derivs.h5"
+                dataset_name = "overlap_deriv" + str(deriv_order)
+            elif os.path.exists("oei_partials.h5"):
+                file_name = "oei_partials.h5"
+                dataset_name = "overlap_deriv" + str(deriv_order) + "_" + str(idx)
+            else:
+                raise Exception("Something went wrong reading integral derivative file")
+            with h5py.File(file_name, 'r') as f:
+                data_set = f[dataset_name]
+                if len(data_set.shape) == 3:
+                    S = data_set[:,:,idx]
+                elif len(data_set.shape) == 2:
+                    S = data_set[:,:]
+                else:
+                    raise Exception("Something went wrong reading integral derivative file")
+            return jnp.asarray(S)
 
-def potential(la,ma,na,lb,mb,nb,aa,bb,PA_pow,PB_pow,Pgeom_pow,boys_eval,prefactor,charges,A_vals):
-    """
-    Computes a single electron-nuclear attraction integral
-    """
-    gamma = aa + bb
-    prefactor *= -2 * jnp.pi / gamma
+    def kinetic_deriv_impl(self, geom, deriv_vec):
+        deriv_vec = np.asarray(deriv_vec, int)
+        deriv_order = np.sum(deriv_vec)
 
-    def loop_val(n, val):
-      Ax = A_array(la,lb,PA_pow[0],PB_pow[0],Pgeom_pow[n,0,:],gamma,A_vals)
-      Ay = A_array(ma,mb,PA_pow[1],PB_pow[1],Pgeom_pow[n,1,:],gamma,A_vals)
-      Az = A_array(na,nb,PA_pow[2],PB_pow[2],Pgeom_pow[n,2,:],gamma,A_vals)
+        #TODO update once core algo in libint is computed, this just computes one slice at a time
+        if self.mode == 'core':
+            T = libint_interface.kinetic_deriv(np.asarray(deriv_vec, int))
+            return jnp.asarray(T).reshape(self.nbf,self.nbf)
+        else:
+            idx = get_deriv_vec_idx(deriv_vec)
+            if os.path.exists("oei_derivs.h5"):
+                file_name = "oei_derivs.h5"
+                dataset_name = "kinetic_deriv" + str(deriv_order)
+            elif os.path.exists("oei_partials.h5"):
+                file_name = "oei_partials.h5"
+                dataset_name = "kinetic_deriv" + str(deriv_order) + "_" + str(idx)
+            else:
+                raise Exception("Something went wrong reading integral derivative file")
+            with h5py.File(file_name, 'r') as f:
+                data_set = f[dataset_name]
+                if len(data_set.shape) == 3:
+                    T = data_set[:,:,idx]
+                elif len(data_set.shape) == 2:
+                    T = data_set[:,:]
+                else:
+                    raise Exception("Something went wrong reading integral derivative file")
+            return jnp.asarray(T)
 
-      I, J, K, total = 0, 0, 0, 0
-      def loop_I(arr0):
-         I_0, J_0, K_0, val_0, total_0 = arr0
-         J_0 = 0
+    def potential_deriv_impl(self, geom, deriv_vec):
+        deriv_vec = np.asarray(deriv_vec, int)
+        deriv_order = np.sum(deriv_vec)
 
-         def loop_J(arr1):
-            I_1, J_1, K_1, val_1, total_1 = arr1
-            K_1 = 0
+        #TODO update once core algo in libint is computed, this just computes one slice at a time
+        if self.mode == 'core':
+            V = libint_interface.potential_deriv(np.asarray(deriv_vec, int))
+            return jnp.asarray(V).reshape(self.nbf,self.nbf)
+        else:
+            idx = get_deriv_vec_idx(deriv_vec)
+            if os.path.exists("oei_derivs.h5"):
+                file_name = "oei_derivs.h5"
+                dataset_name = "potential_deriv" + str(deriv_order)
+            elif os.path.exists("oei_partials.h5"):
+                file_name = "oei_partials.h5"
+                dataset_name = "potential_deriv" + str(deriv_order) + "_" + str(idx)
+            else:
+                raise Exception("Something went wrong reading integral derivative file")
+            with h5py.File(file_name, 'r') as f:
+                data_set = f[dataset_name]
+                if len(data_set.shape) == 3:
+                    V = data_set[:,:,idx]
+                elif len(data_set.shape) == 2:
+                    V = data_set[:,:]
+                else:
+                    raise Exception("Something went wrong reading integral derivative file")
+            return jnp.asarray(V)
 
-            def loop_K(arr2):
-               I_2, J_2, K_2, val_2, total_2 = arr2
-               total_2 += Ax[I_2] * Ay[J_2] * Az[K_2] * boys_eval[I_2 + J_2 + K_2, n]
-               K_2 += 1
-               return (I_2, J_2, K_2, val_2, total_2)
+    def overlap_jvp(self, primals, tangents):
+        geom, = primals
+        primals_out = self.overlap(geom)
+        tangents_out = self.overlap_deriv(geom, tangents[0])
+        return primals_out, tangents_out
 
-            I_1_, J_1_, K_1_, val_1_, total_1_ = while_loop(lambda arr2: arr2[2] < na + nb + 1, loop_K, (I_1, J_1, K_1, val_1, total_1))
-            J_1_ += 1
-            return (I_1_, J_1_, K_1_, val_1_, total_1_)
+    def overlap_deriv_jvp(self, primals, tangents):
+        geom, deriv_vec = primals
+        primals_out = self.overlap_deriv(geom, deriv_vec)
+        tangents_out = self.overlap_deriv(geom, deriv_vec + tangents[0])
+        return primals_out, tangents_out
 
-         I_0_, J_0_, K_0_, val_0_, total_0_ = while_loop(lambda arr1: arr1[1] < ma + mb + 1, loop_J, (I_0, J_0, K_0, val_0, total_0))
-         I_0_ += 1
-         return (I_0_, J_0_, K_0_, val_0_, total_0_)
+    def kinetic_jvp(self, primals, tangents):
+        geom, = primals
+        primals_out = self.kinetic(geom)
+        tangents_out = self.kinetic_deriv(geom, tangents[0])
+        return primals_out, tangents_out
 
-      I_, J_, K_, val_, total_ = while_loop(lambda arr0: arr0[0] < la + lb + 1, loop_I, (I, J, K, val, total))
-      val_ += charges[n] * prefactor * total_
-      return val_
+    def kinetic_deriv_jvp(self, primals, tangents):
+        geom, deriv_vec = primals
+        primals_out = self.kinetic_deriv(geom, deriv_vec)
+        tangents_out = self.kinetic_deriv(geom, deriv_vec + tangents[0])
+        return primals_out, tangents_out
 
-    val = fori_loop(0, Pgeom_pow.shape[0], loop_val, 0)
-    return val
+    def potential_jvp(self, primals, tangents):
+        geom, = primals
+        primals_out = self.potential(geom)
+        tangents_out = self.potential_deriv(geom, tangents[0])
+        return primals_out, tangents_out
 
-def oei_arrays(geom, basis, charges):
-    """
-    Build one electron integral arrays (overlap, kinetic, and potential integrals)
-    """
-    coeffs, exps, atoms, ams, indices, dims = flatten_basis_data(basis)
-    nbf = get_nbf(basis)
-    nprim = coeffs.shape[0]
-    max_am = jnp.max(ams)
-    A_vals = jnp.zeros(2*max_am+1)
+    def potential_deriv_jvp(self, primals, tangents):
+        geom, deriv_vec = primals
+        primals_out = self.potential_deriv(geom, deriv_vec)
+        tangents_out = self.potential_deriv(geom, deriv_vec + tangents[0])
+        return primals_out, tangents_out
 
-    # Save various AM distributions for indexing
-    # Obtain all possible primitive quartet index combinations 
-    primitive_duets = cartesian_product(jnp.arange(nprim), jnp.arange(nprim))
-    STV = jnp.zeros((3,nbf,nbf))
+    # Define Batching rules, this is only needed since jax.jacfwd will call vmap on the JVP's
+    # of each oei function
+    def overlap_deriv_batch(self, batched_args, batch_dims):
+        # When the input argument of deriv_batch is batched along the 0'th axis
+        # we want to evaluate every 2d slice, gather up a (ncart, n,n) array,
+        # (expand dims at 0 and concatenate at 0)
+        # and then return the results, indicating the out batch axis
+        # is in the 0th position (return results, 0)
+        geom_batch, deriv_batch = batched_args
+        geom_dim, deriv_dim = batch_dims
+        results = []
+        for i in deriv_batch:
+            tmp = self.overlap_deriv(geom_batch, i)
+            results.append(jnp.expand_dims(tmp, axis=0))
+        results = jnp.concatenate(results, axis=0)
+        return results, 0
 
-    for n in range(primitive_duets.shape[0]):
-       p1,p2 = primitive_duets[n]
-       coef = coeffs[p1] * coeffs[p2]
-       aa, bb = exps[p1], exps[p2]
-       atom1, atom2 = atoms[p1], atoms[p2]
-       am1, am2 = ams[p1], ams[p2]
-       A, B = geom[atom1], geom[atom2]
-       ld1, ld2 = am_leading_indices[am1], am_leading_indices[am2]
+    def kinetic_deriv_batch(self, batched_args, batch_dims):
+        geom_batch, deriv_batch = batched_args
+        geom_dim, deriv_dim = batch_dims
+        results = []
+        for i in deriv_batch:
+            tmp = self.kinetic_deriv(geom_batch, i)
+            results.append(jnp.expand_dims(tmp, axis=0))
+        results = jnp.concatenate(results, axis=0)
+        return results, 0
 
-       gamma = aa + bb
-       prefactor = jnp.exp(-aa * bb * jnp.dot(A-B,A-B) / gamma)
-       P = (aa * A + bb * B) / gamma
-       # Maximum angular momentum: hard coded
-       #max_am = 3 # f function support
-       # Precompute all powers up to 2+max_am of Pi-Ai, Pi-Bi.
-       # We need 2+max_am since kinetic requires incrementing angluar momentum by +2
-       PA_pow = jnp.power(jnp.broadcast_to(P-A, (max_am+3, 3)).T, jnp.arange(max_am+3))
-       PB_pow = jnp.power(jnp.broadcast_to(P-B, (max_am+3, 3)).T, jnp.arange(max_am+3))
-
-       # For potential integrals, we need the difference between
-       # the gaussian product center P and ALL atoms in the molecule,
-       # and then take all possible powers up to 2*max_am.
-       # We pre-collect this into a 3d array, and then just pull out what we need via indexing in the loops, so they need not be recomputed.
-       # The resulting array has dimensions (atom, cartesian component, power) so index (0, 1, 3) would return (Py - atom0_y)^3
-       P_minus_geom = jnp.broadcast_to(P, geom.shape) - geom
-       Pgeom_pow = jnp.power(jnp.transpose(jnp.broadcast_to(P_minus_geom, (2*max_am + 1,geom.shape[0],geom.shape[1])), (1,2,0)), jnp.arange(2*max_am + 1))
-       # All possible jnp.dot(P-atom,P-atom)
-       rcp2 = jnp.einsum('ij,ij->i', P_minus_geom, P_minus_geom)
-       # All needed (and unneeded, for am < max_am) boys function evaluations
-       boys_arg = jnp.broadcast_to(rcp2 * gamma, (2*max_am+1, geom.shape[0]))
-       boys_nu = jnp.tile(jnp.arange(2*max_am+1), (geom.shape[0],1)).T
-       boys_eval = boys(boys_nu,boys_arg)
-
-       a, b = 0, 0
-       def loop_a(arr0):
-          a_0, b_0, oei_0 = arr0
-          b_0 = 0
-
-          def loop_b(arr1):
-             a_1, b_1, oei_1 = arr1
-             # Gather angular momentum and index
-             la,ma,na = angular_momentum_combinations[a_1 + ld1]
-             lb,mb,nb = angular_momentum_combinations[b_1 + ld2]
-             # To only create unique indices, need to have separate indices arrays for i and j.
-             i = indices[p1] + a_1
-             j = indices[p2] + b_1
-             # Compute one electron integrals and add to appropriate index
-             overlap_int = overlap(la,ma,na,lb,mb,nb,aa,bb,PA_pow,PB_pow,prefactor) * coef
-             kinetic_int = kinetic(la,ma,na,lb,mb,nb,aa,bb,PA_pow,PB_pow,prefactor) * coef
-             potential_int = potential(la,ma,na,lb,mb,nb,aa,bb,PA_pow,PB_pow,Pgeom_pow,boys_eval,prefactor,charges,A_vals) * coef
-             oei_1 = oei_1.at[([0,1,2],[i,i,i],[j,j,j])].set((overlap_int, kinetic_int, potential_int))
-             b_1 += 1
-             return (a_1, b_1, oei_1)
-
-          a_0_, b_0_, oei_0_ = while_loop(lambda arr1: arr1[1] < dims[p2], loop_b, (a_0, b_0, oei_0))
-          a_0_ += 1
-          return (a_0_, b_0_, oei_0_)
-
-       a_, b_, oei_ = while_loop(lambda arr0: arr0[0] < dims[p1], loop_a, (a, b, STV))
-
-       return oei_
-
-    return STV[0], STV[1], STV[2]
+    def potential_deriv_batch(self, batched_args, batch_dims):
+        geom_batch, deriv_batch = batched_args
+        geom_dim, deriv_dim = batch_dims
+        results = []
+        for i in deriv_batch:
+            tmp = self.potential_deriv(geom_batch, i)
+            results.append(jnp.expand_dims(tmp, axis=0))
+        results = jnp.concatenate(results, axis=0)
+        return results, 0
