+++ quax/methods/ccsd.py
@@ -27,8 +27,8 @@ def rccsd(geom, basis_name, xyz_path, nuclear_charges, charge, options, deriv_or
     fock_Vd = eps[v]
 
     # Oribital energy denominators 
-    D = 1.0 / (fock_Od.reshape(-1,1,1,1) + fock_Od.reshape(-1,1,1) - fock_Vd.reshape(-1,1) - fock_Vd)
-    d = 1.0 / (fock_Od.reshape(-1,1) - fock_Vd)
+    D = 1.0 / (fock_Od.reshape(-1, 1, 1, 1) + fock_Od.reshape(-1, 1, 1) - fock_Vd.reshape(-1, 1) - fock_Vd)
+    d = 1.0 / (fock_Od.reshape(-1, 1) - fock_Vd)
 
     # Initial Amplitudes
     T1 = jnp.zeros((ndocc,nvir))
@@ -42,7 +42,7 @@ def rccsd(geom, basis_name, xyz_path, nuclear_charges, charge, options, deriv_or
         E_old = E_ccsd * 1
 
         T1, T2 = rccsd_iter(T1, T2, V, d, D, ndocc, nvir)
-        E_ccsd = rccsd_energy(T1,T2,V[2])
+        E_ccsd = rccsd_energy(T1, T2, V[2])
 
         iteration += 1
         if iteration == maxit:
@@ -60,10 +60,10 @@ def rccsd(geom, basis_name, xyz_path, nuclear_charges, charge, options, deriv_or
 @jax.jit
 def rccsd_energy(T1, T2, Voovv):
     E_ccsd = 0.0
-    E_ccsd -= jnp.tensordot(T1, jnp.tensordot(T1, Voovv, [(0,1),(1,2)]), [(0,1),(0,1)])
-    E_ccsd -= jnp.tensordot(T2, Voovv, [(0,1,2,3),(1,0,2,3)])
-    E_ccsd += 2.0*jnp.tensordot(T2, Voovv, [(0,1,2,3),(0,1,2,3)])
-    E_ccsd += 2.0*jnp.tensordot(T1, jnp.tensordot(T1, Voovv, [(0,1),(0,2)]), [(0,1),(0,1)])
+    E_ccsd -= jnp.tensordot(T1, jnp.tensordot(T1, Voovv, [(0, 1), (1, 2)]), [(0, 1), (0, 1)])
+    E_ccsd -= jnp.tensordot(T2, Voovv, [(0, 1, 2, 3), (1, 0, 2, 3)])
+    E_ccsd += 2.0*jnp.tensordot(T2, Voovv, [(0, 1, 2, 3),(0, 1, 2, 3)])
+    E_ccsd += 2.0*jnp.tensordot(T1, jnp.tensordot(T1, Voovv, [(0, 1), (0, 2)]), [(0, 1), (0, 1)])
     return E_ccsd
 
 # Jit compiling ccsd is a BAD IDEA.
@@ -75,9 +75,9 @@ def rccsd_iter(T1, T2, V, d, D, ndocc, nvir):
     newT2 = jnp.zeros(T2.shape)
 
     # T1 equation
-    newT1 += jnp.tensordot(T1, Voovv, [(0,1),(0,2)])
-    newT1 += jnp.tensordot(T2, Vovvv, [(1,2,3), (0,3,2)])
-    newT1 -= jnp.tensordot(Vooov, T2, [(0,1,3),(0,1,3)])
+    newT1 += jnp.tensordot(T1, Voovv, [(0, 1), (0, 2)])
+    newT1 += jnp.tensordot(T2, Vovvv, [(1, 2, 3), (0, 3, 2)])
+    newT1 -= jnp.tensordot(Vooov, T2, [(0, 1, 3), (0, 1, 3)])
     newT1 -= jnp.einsum('kc, la, lkic -> ia', T1, T1, Vooov, optimize = 'optimal')
     newT1 += jnp.einsum('kc, id, kacd -> ia', T1, T1, Vovvv, optimize = 'optimal')
     newT1 -= jnp.einsum('kc, ilad, lkcd -> ia', T1, T2, Voovv, optimize = 'optimal')
@@ -85,12 +85,12 @@ def rccsd_iter(T1, T2, V, d, D, ndocc, nvir):
     newT1 -= jnp.einsum('ic, lkad, lkcd -> ia', T1, T2, Voovv, optimize = 'optimal')
     newT1 -= jnp.einsum('la, ikdc, klcd -> ia', T1, T2, Voovv, optimize = 'optimal')
     newT1 -= jnp.einsum('kc, id, la, klcd -> ia', T1, T1, T1, Voovv, optimize = 'optimal')
-    newT1 += 2.0*jnp.einsum('kc, ilad, klcd -> ia', T1, T2, Voovv, optimize = 'optimal')
+    newT1 += 2.0 * jnp.einsum('kc, ilad, klcd -> ia', T1, T2, Voovv, optimize = 'optimal')
     newT1 *= 2.0
 
-    newT1 -= jnp.tensordot(T1, Vovov, [(0,1),(2,1)])
-    newT1 -= jnp.tensordot(T2, Vovvv, [(0,2,3),(0,3,2)])
-    newT1 += jnp.tensordot(Vooov, T2, [(0,1,3),(1,0,3)])
+    newT1 -= jnp.tensordot(T1, Vovov, [(0, 1), (2, 1)])
+    newT1 -= jnp.tensordot(T2, Vovvv, [(0, 2, 3), (0, 3, 2)])
+    newT1 += jnp.tensordot(Vooov, T2, [(0, 1, 3), (1, 0, 3)])
     newT1 -= jnp.einsum('kc, id, kadc -> ia', T1, T1, Vovvv, optimize = 'optimal')
     newT1 += jnp.einsum('kc, la, klic -> ia', T1, T1, Vooov, optimize = 'optimal')
     newT1 += jnp.einsum('kc, liad, lkcd -> ia', T1, T2, Voovv, optimize = 'optimal')
@@ -104,16 +104,16 @@ def rccsd_iter(T1, T2, V, d, D, ndocc, nvir):
     newT2 -= jnp.einsum('ikac, jlbd, lkcd -> ijab', T2, T2, Voovv, optimize = 'optimal')
     newT2 -= jnp.einsum('kiac, jlbd, klcd -> ijab', T2, T2, Voovv, optimize = 'optimal')
     newT2 -= jnp.einsum('ijac, klbd, klcd -> ijab', T2, T2, Voovv, optimize = 'optimal')
-    newT2 += 2.0*jnp.einsum('ikac, jlbd, klcd -> ijab', T2, T2, Voovv, optimize = 'optimal')
+    newT2 += 2.0 * jnp.einsum('ikac, jlbd, klcd -> ijab', T2, T2, Voovv, optimize = 'optimal')
     newT2 *= 2.0
 
     # Reducing Vvvvv contractions to tensordot is especially productive.
     # TODO try reducing Vovvv as well. Also check if removing jit makes this optimization moot...
     newT2 += Voovv
-    newT2 += jnp.tensordot(T1, jnp.tensordot(T1, Vvvvv, [(1,),(1,)]), [(1,),(1,)])
-    newT2 += jnp.tensordot(T2, Vvvvv, [(2,3),(0,1)])
+    newT2 += jnp.tensordot(T1, jnp.tensordot(T1, Vvvvv, [(1, ), (1, )]), [(1, ), (1, )])
+    newT2 += jnp.tensordot(T2, Vvvvv, [(2, 3), (0, 1)])
     newT2 += jnp.einsum('ka, lb, ijkl -> ijab', T1, T1, Voooo, optimize = 'optimal')
-    newT2 += jnp.tensordot(T2, Voooo, [(0,1),(2,3)]).transpose((2,3,0,1))
+    newT2 += jnp.tensordot(T2, Voooo, [(0, 1), (2, 3)]).transpose((2, 3, 0, 1))
     newT2 -= jnp.einsum('ic, jd, ka, kbcd -> ijab', T1, T1, T1, Vovvv, optimize = 'optimal')
     newT2 -= jnp.einsum('ic, jd, kb, kadc -> ijab', T1, T1, T1, Vovvv, optimize = 'optimal')
     newT2 += jnp.einsum('ic, ka, lb, lkjc -> ijab', T1, T1, T1, Vooov, optimize = 'optimal')
@@ -129,7 +129,7 @@ def rccsd_iter(T1, T2, V, d, D, ndocc, nvir):
     newT2 += jnp.einsum('ic, jd, lkab, lkcd -> ijab', T1, T1, T2, Voovv, optimize = 'optimal')
     newT2 += jnp.einsum('ka, lb, ijdc, lkcd -> ijab', T1, T1, T2, Voovv, optimize = 'optimal')
 
-    P_OVVO  = jnp.tensordot(T2, Voovv, [(1,3),(0,2)]).transpose((0,2,1,3))
+    P_OVVO  = jnp.tensordot(T2, Voovv, [(1, 3),(0, 2)]).transpose((0, 2, 1, 3))
     P_OVVO -= jnp.einsum('lb, ikac, lkjc -> ijab', T1, T2, Vooov, optimize = 'optimal')
     P_OVVO += jnp.einsum('jc, ikad, kbdc -> ijab', T1, T2, Vovvv, optimize = 'optimal')
     P_OVVO += jnp.einsum('kc, ijad, kbcd -> ijab', T1, T2, Vovvv, optimize = 'optimal')
@@ -140,13 +140,13 @@ def rccsd_iter(T1, T2, V, d, D, ndocc, nvir):
     P_OVVO -= jnp.einsum('ikdc, ljab, klcd -> ijab', T2, T2, Voovv, optimize = 'optimal')
     P_OVVO *= 2.0
 
-    P_OVVO -= jnp.tensordot(T1, Vooov, [(0,), (2,)]).transpose((2,1,3,0))
-    P_OVVO += jnp.tensordot(T1, Vovvv, [(1,),(1,)]).transpose((1,0,2,3))
-    P_OVVO -= jnp.tensordot(T2, Voovv, [(0,3),(0,2)]).transpose((0,2,1,3))
+    P_OVVO -= jnp.tensordot(T1, Vooov, [(0, ), (2, )]).transpose((2, 1, 3, 0))
+    P_OVVO += jnp.tensordot(T1, Vovvv, [(1, ), (1, )]).transpose((1, 0, 2, 3))
+    P_OVVO -= jnp.tensordot(T2, Voovv, [(0, 3), (0, 2)]).transpose((0, 2, 1, 3))
     P_OVVO -= jnp.einsum('ic, ka, kjcb -> ijab', T1, T1, Voovv, optimize = 'optimal')
     P_OVVO -= jnp.einsum('ic, kb, jcka -> ijab', T1, T1, Vovov, optimize = 'optimal')
-    P_OVVO -= jnp.tensordot(T2, Vovov, [(1,3),(2,1)]).transpose((0,2,1,3))
-    P_OVVO -= jnp.tensordot(T2, Vovov, [(0,3),(2,1)]).transpose((2,0,1,3))
+    P_OVVO -= jnp.tensordot(T2, Vovov, [(1, 3), (2, 1)]).transpose((0, 2, 1, 3))
+    P_OVVO -= jnp.tensordot(T2, Vovov, [(0, 3), (2, 1)]).transpose((2, 0, 1, 3))
     P_OVVO += jnp.einsum('lb, kiac, lkjc -> ijab', T1, T2, Vooov, optimize = 'optimal')
     P_OVVO -= jnp.einsum('jc, ikdb, kacd -> ijab', T1, T2, Vovvv, optimize = 'optimal')
     P_OVVO -= jnp.einsum('jc, kiad, kbdc -> ijab', T1, T2, Vovvv, optimize = 'optimal')
@@ -165,7 +165,7 @@ def rccsd_iter(T1, T2, V, d, D, ndocc, nvir):
     P_OVVO += jnp.einsum('ic, lb, kjad, klcd -> ijab', T1, T1, T2, Voovv, optimize = 'optimal')
 
     newT2 += P_OVVO 
-    newT2 += P_OVVO.transpose((1,0,3,2))
+    newT2 += P_OVVO.transpose((1, 0, 3, 2))
 
     newT1 *= d
     newT2 *= D
